Basic dotfiles test
It turned out the argument I had used for get_host was causing the
tests to be run against my local machine, rather than the Vagrant box.
Sadly, the hard coded connection details was the only way I could get
this working. When I tried to use ssh_config, either from the command
line or passing it as an argument to get_host, it just wouldn't use
the connection. This will have to do for now.