Add target for 'corporate mode'
This is for when we're behind a proxy in some kind of corporate
environment. It's not working as planned at the moment because I
couldn't figure out how to get Make to pass the necessary environment
variables, so right now it's relying on an environment variable being
set before it runs.